♻️ Move OpenOptions and DefaultOpenOptions to opts
